# Add DoWithData function