👋 I am not sure that you will consider this PR, but it is to add a `DoWithData` generic function that can also return data from a retry, not just an error. That is, it resolves #58. Of course, with this change I had to bump to Go 1.18.

If you think this is a useful feature, I could add some more tests and documentation.

I wasn't sure what to name this function so I borrowed the idea from https://pkg.go.dev/github.com/cenkalti/backoff/v4, which has a `RetryWithData` function which does a similar thing as this.

Thanks for the very useful library ❤️