@craigpastro 2023-04-22
Hi @JaSei 👋 That's great. It will be nice to get this functionality here. I think you are right about supporting only 1.18+. 

I am thinking about what to do with the tests. I could create two of each for `Do` and `DoWithData`. Or I could rewrite the majority of them using `DoWithData` as `Do` uses `DoWithData` under the hood. WDYT?