@JaSei 2023-08-03
I'm just afraid of performance, because generics affects performance. That will be awesome to do benchmark before and with this change 🤔