@JaSei 2023-08-03
I've just merged #89  where some benchmark are available