@craigpastro 2023-08-03
I added similar `DoWithData` benchmarks. The results are
```
$ go test -benchmem -run=^$ -bench ^BenchmarkDo github.com/avast/retry-go/v4
goos: darwin
goarch: arm64
pkg: github.com/avast/retry-go/v4
BenchmarkDo-12                                 3         464051111 ns/op            2698 B/op         48 allocs/op
BenchmarkDoWithData-12                         3         433661403 ns/op            2725 B/op         47 allocs/op
BenchmarkDoNoErrors-12                  12527132                96.55 ns/op          208 B/op          4 allocs/op
BenchmarkDoWithDataNoErrors-12          12521830                94.29 ns/op          208 B/op          4 allocs/op
PASS
ok      github.com/avast/retry-go/v4    9.301s
```
So the generic version is actually faster on my computer 🤷