@JaSei 2023-08-04
`Do` function just wrap `DoWithData` right? It's not strange it is slower.