@JaSei 2023-08-04
current

```
go test -count 10 -benchmem -run=^$ -bench ^BenchmarkDo github.com/avast/retry-go/v4 | tee current.txt
goos: darwin
goarch: amd64
pkg: github.com/avast/retry-go/v4
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkDo-16            	       3	 474128987 ns/op	    2730 B/op	      48 allocs/op
BenchmarkDo-16            	       3	 441499631 ns/op	    2725 B/op	      47 allocs/op
BenchmarkDo-16            	       3	 449390845 ns/op	    2693 B/op	      47 allocs/op
BenchmarkDo-16            	       3	 488695333 ns/op	    2725 B/op	      47 allocs/op
BenchmarkDo-16            	       2	 601685067 ns/op	    2704 B/op	      48 allocs/op
BenchmarkDo-16            	       3	 336872997 ns/op	    2693 B/op	      47 allocs/op
BenchmarkDo-16            	       3	 384347911 ns/op	    2725 B/op	      47 allocs/op
BenchmarkDo-16            	       3	 480906307 ns/op	    2693 B/op	      47 allocs/op
BenchmarkDo-16            	       3	 455362447 ns/op	    2693 B/op	      47 allocs/op
BenchmarkDo-16            	       3	 443170384 ns/op	    2693 B/op	      47 allocs/op
BenchmarkDoNoErrors-16    	 6872852	       159.4 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoNoErrors-16    	 7650360	       161.3 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoNoErrors-16    	 7235683	       159.3 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoNoErrors-16    	 7465636	       160.2 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoNoErrors-16    	 7549692	       160.7 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoNoErrors-16    	 7510610	       159.8 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoNoErrors-16    	 7438124	       160.3 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoNoErrors-16    	 7416504	       160.2 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoNoErrors-16    	 7356183	       160.4 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoNoErrors-16    	 7393480	       160.1 ns/op	     208 B/op	       4 allocs/op
PASS
ok  	github.com/avast/retry-go/v4	35.971s
```

generic
```
go test -count 10 -benchmem -run=^$ -bench ^BenchmarkDo github.com/avast/retry-go/v4 | tee generic.txt
goos: darwin
goarch: amd64
pkg: github.com/avast/retry-go/v4
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkDo-16                    	       3	 406306609 ns/op	    2701 B/op	      48 allocs/op
BenchmarkDo-16                    	       3	 419470846 ns/op	    2693 B/op	      47 allocs/op
BenchmarkDo-16                    	       2	 567716303 ns/op	    2696 B/op	      47 allocs/op
BenchmarkDo-16                    	       2	 562713288 ns/op	    2696 B/op	      47 allocs/op
BenchmarkDo-16                    	       3	 418301987 ns/op	    2693 B/op	      47 allocs/op
BenchmarkDo-16                    	       2	 541207332 ns/op	    2696 B/op	      47 allocs/op
BenchmarkDo-16                    	       2	 526211617 ns/op	    2696 B/op	      47 allocs/op
BenchmarkDo-16                    	       2	 517419526 ns/op	    2696 B/op	      47 allocs/op
BenchmarkDo-16                    	       3	 478391497 ns/op	    2693 B/op	      47 allocs/op
BenchmarkDo-16                    	       3	 452548175 ns/op	    2725 B/op	      47 allocs/op
BenchmarkDoWithData-16            	       3	 463040866 ns/op	    2693 B/op	      47 allocs/op
BenchmarkDoWithData-16            	       3	 496158943 ns/op	    2693 B/op	      47 allocs/op
BenchmarkDoWithData-16            	       3	 488367012 ns/op	    2725 B/op	      47 allocs/op
BenchmarkDoWithData-16            	       3	 454618897 ns/op	    2693 B/op	      47 allocs/op
BenchmarkDoWithData-16            	       3	 435430056 ns/op	    2693 B/op	      47 allocs/op
BenchmarkDoWithData-16            	       2	 552289967 ns/op	    2744 B/op	      48 allocs/op
BenchmarkDoWithData-16            	       3	 569748815 ns/op	    2693 B/op	      47 allocs/op
BenchmarkDoWithData-16            	       3	 416597207 ns/op	    2725 B/op	      47 allocs/op
BenchmarkDoWithData-16            	       3	 358455415 ns/op	    2725 B/op	      47 allocs/op
BenchmarkDoWithData-16            	       3	 455297803 ns/op	    2725 B/op	      47 allocs/op
BenchmarkDoNoErrors-16            	 7035135	       161.9 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoNoErrors-16            	 7389806	       161.3 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoNoErrors-16            	 7394016	       161.5 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoNoErrors-16            	 7380039	       162.2 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoNoErrors-16            	 7424865	       162.2 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoNoErrors-16            	 7111860	       160.5 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoNoErrors-16            	 7285305	       162.6 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoNoErrors-16            	 7410627	       160.7 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoNoErrors-16            	 7340961	       161.6 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoNoErrors-16            	 7295727	       164.1 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoWithDataNoErrors-16    	 7357304	       159.9 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoWithDataNoErrors-16    	 6649852	       166.9 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoWithDataNoErrors-16    	 6938404	       176.3 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoWithDataNoErrors-16    	 7181965	       160.4 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoWithDataNoErrors-16    	 7311484	       166.2 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoWithDataNoErrors-16    	 6939157	       169.7 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoWithDataNoErrors-16    	 6648344	       179.0 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoWithDataNoErrors-16    	 6794847	       177.0 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoWithDataNoErrors-16    	 6782588	       171.4 ns/op	     208 B/op	       4 allocs/op
BenchmarkDoWithDataNoErrors-16    	 7279119	       166.9 ns/op	     208 B/op	       4 allocs/op
PASS
ok  	github.com/avast/retry-go/v4	73.128s
```

benchstat
```
benchstat current.txt generic.txt
goos: darwin
goarch: amd64
pkg: github.com/avast/retry-go/v4
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
                      │ current.txt  │             generic.txt             │
                      │    sec/op    │    sec/op     vs base               │
Do-16                   452.4m ± 15%   497.9m ± 16%       ~ (p=0.353 n=10)
DoNoErrors-16           160.2n ±  0%   161.8n ±  1%  +0.97% (p=0.000 n=10)
DoWithData-16                          459.2m ± 20%
DoWithDataNoErrors-16                  168.3n ±  5%
geomean                 269.2µ         280.9µ        +5.42%

                      │ current.txt  │              generic.txt              │
                      │     B/op     │     B/op      vs base                 │
Do-16                   2.635Ki ± 1%   2.633Ki ± 0%       ~ (p=0.745 n=10)
DoNoErrors-16             208.0 ± 0%     208.0 ± 0%       ~ (p=1.000 n=10) ¹
DoWithData-16                          2.646Ki ± 1%
DoWithDataNoErrors-16                    208.0 ± 0%
geomean                   749.2          749.7       -0.05%
¹ all samples are equal

                      │ current.txt │             generic.txt             │
                      │  allocs/op  │ allocs/op   vs base                 │
Do-16                    47.00 ± 2%   47.00 ± 0%       ~ (p=1.000 n=10)
DoNoErrors-16            4.000 ± 0%   4.000 ± 0%       ~ (p=1.000 n=10) ¹
DoWithData-16                         47.00 ± 0%
DoWithDataNoErrors-16                 4.000 ± 0%
geomean                  13.71        13.71       +0.00%
¹ all samples are equal
```


`DoNoErrors` is only about 1% slower - that is a good result I thing.